Add !do and !eval owner commands
The !do command runs python code. Can be formatted with discord code.
Use the say(msg) function to have the bot print a message. Keep in mind that it will not yield when printing, so it moves on immediately. There is no way to sleep in this code.

The !eval command evaluates anything and gives the result. This would normally be expressions such as 2 * 3 + 4