Ability to assert expressions in commands and reply with the error message.
The ability to assert is particularly useful as it takes less space and is simple to understand. If you do not specify an error message, the bot will look for command.error. If this is also unspecified, it will simply show the command help.

Renamed "sub command" in comments and docstrings to "subcommand"