Add purge command.
Renamed client.last_deleted_message to client.last_deleted_messages, and change functionality to check if a message is actually already in the last deleted messages.

Overwrite client.delete_messages to add the deleted messages in bulk to the last deleted, in order to prevent the bot from posting anything in the changelog.